fix some mistakes (constant name change affected target of bullet :/)

+++ src/main/scala/Logic/GameObject.scala
@@ -4,7 +4,7 @@ package Logic
 
 import scalafx.scene.image.{ImageView, Image}
 
-trait GameObject(path: String, size: Int) extends ImageView {
+trait GameObject(path: String, size: Double) extends ImageView {
   image = new Image(path)
   fitWidth = size
   fitHeight = size
